+++ model_bakery/generators.py
@@ -80,10 +80,12 @@ try:
     from django.contrib.postgres.fields.ranges import (
         IntegerRangeField,
         BigIntegerRangeField,
+        FloatRangeField,
     )
 except ImportError:
     IntegerRangeField = None
     BigIntegerRangeField = None
+    FloatRangeField = None
 
 
 def _make_integer_gen_by_range(field_type):
@@ -163,6 +165,8 @@ if IntegerRangeField:
     default_mapping[IntegerRangeField] = _make_pg_numbers_range(int)
 if BigIntegerRangeField:
     default_mapping[BigIntegerRangeField] = _make_pg_numbers_range(int)
+if FloatRangeField:
+    default_mapping[FloatRangeField] = _make_pg_numbers_range(float)
 
 
 # Add GIS fields